requirements.txt
python-telegram-bot[webhooks]==22.1
# остальные библиотеки — как есть
imap-tools
apscheduler
pandas
openpyxl
psycopg2-binary
requests>=2.31.0
python-dotenv
sqlalchemy
alembic
